% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARA.R
\name{ARA}
\alias{ARA}
\title{ARA}
\usage{
ARA(
  label = "ARA",
  dict = mpipoet::mpipoet_dict,
  arrange_vertically = "auto",
  button_style = "min-width: 290px",
  ...
)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{arrange_vertically}{(Boolean scalar) Flag whether to arrange buttons vertically.}

\item{button_style}{(Character scalar) CSS style for buttons}

\item{...}{Further arguments to be passed to \code{\link{ARA}()}.}
}
\description{
This function defines a ARA module for incorporation into a
psychTestR timeline.
Use this function if you want to include the Aesthetic Reponsiveness Assessment (AReA) in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the ARA,
consider using \code{\link{ARA_standalone}()}.
}
