% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{ARA_standalone}
\alias{ARA_standalone}
\title{ARA Standalone}
\usage{
ARA_standalone(languages = mpipoet::languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), and \code{"de"} (German).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{ARA_standalone}()}.}
}
\description{
This function launches a standalone testing session for the ARA questionnaire.
ARA stands for 'Aesthetic Responsiveness Assessment (AReA)'.
}
