% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{ART}
\alias{ART}
\title{ART}
\usage{
ART(
  num_items = NULL,
  mode = "pairs",
  with_welcome = TRUE,
  with_finish = TRUE,
  with_feedback = FALSE,
  label = "ART",
  dict = mpipoet::mpipoet_dict,
  timeout = 10,
  ...
)
}
\arguments{
\item{num_items}{(Integer scalar) Number of items in the test. Default NULL pulls all items.}

\item{mode}{(String scalar) Presentation mode of the ART, 'single', 'pairs' or 'single_page'. Default is 'pairs'.}

\item{with_welcome}{(Logical scalar) Whether to show a welcome page.}

\item{with_finish}{(Logical scalar) Whether to show a finished page.}

\item{with_feedback}{(Logical scalar) Whether to include feedback to the participants.}

\item{label}{(Character scalar) Label to give the ART results in the output file.}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{timeout}{(Double scalar) The time to answer (in seconds)}

\item{...}{Further arguments to be passed to \code{ART_main_test()}.}
}
\description{
This function defines a ART  module for incorporation into a
psychTestR timeline.
Use this function if you want to include the MSM in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For demoing the MSM, consider using \code{\link{ART_demo}()}.
For a standalone implementation of the ART,
consider using \code{\link{ART_standalone}()}.
}
