% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{ART_demo}
\alias{ART_demo}
\title{Demo ART}
\usage{
ART_demo(
  num_items = 3L,
  mode = "pairs",
  timeout = ifelse(mode == "single_page", 180, 10),
  title = "ART Demo",
  dict = mpipoet::mpipoet_dict,
  admin_password = "demo",
  researcher_email = "klaus.frieler@ae.mpg.de",
  language = c("en", "de", "de_f")
)
}
\arguments{
\item{num_items}{(Integer scalar) Number of items in the test. Default NULL pulls all items.}

\item{mode}{(String scalar) Presentation mode of the ART, 'single', 'pairs' or 'single_page'. Default is 'pairs'.}

\item{timeout}{(Double scalar) The time to answer (in seconds)}

\item{title}{(Character scalar) The title}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.
Defaults to \code{"demo"}.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.
Defaults to \email{longgold@gold.uc.ak},
the email address of this package's developer.}

\item{language}{The language you want to run your demo in.
Possible languages include English (\code{"en"}) and German (informal: \code{"de"}, formal: \code{"de_f"}).
The first language is selected by default}

\item{...}{Further arguments to be passed to \code{\link{ART}()}.}
}
\description{
This function launches a demo for the ART
}
