% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{ART_standalone}
\alias{ART_standalone}
\title{Standalone ART}
\usage{
ART_standalone(
  title = NULL,
  num_items = NULL,
  mode = "pairs",
  timeout = ifelse(mode == "pairs", 10, 120),
  with_id = FALSE,
  with_welcome = TRUE,
  with_feedback = TRUE,
  admin_password = "conifer",
  researcher_email = "klaus.frieler@ae.mpg.de",
  languages = c("en", "de"),
  dict = mpipoet::mpipoet_dict,
  validate_id = "auto",
  ...
)
}
\arguments{
\item{title}{(Scalar character) Title to display during testing.}

\item{num_items}{(Scalar integer) Number of items to be adminstered. Default NULL pulls all items.}

\item{mode}{(String scalar) Presentation mode of the ART, 'single', 'pairs' or 'single_page'. Default is 'pairs'.}

\item{timeout}{(Double scalar) The time to answer (in seconds)}

\item{with_id}{(Logical scalar) Whether to show a ID page.}

\item{with_welcome}{(Logical scalar) Whether to show a welcome page.}

\item{with_feedback}{(Logical scalar) Whether to include feedback to the participants.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}),
and German (\code{"DE"}).
The first language is selected by default}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{validate_id}{(Character scalar or closure) Function for validating IDs or string "auto" for default validation
which means ID should consist only of  alphanumeric characters.}

\item{...}{Further arguments to be passed to \code{\link{ART}()}.}
}
\description{
This function launches a standalone testing session for the ART
This can be used for data collection, either in the laboratory or online.
}
