% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFA.R
\name{BFA}
\alias{BFA}
\title{BFA}
\usage{
BFA(
  label = "BFA",
  dict = mpipoet::mpipoet_dict,
  arrange_vertically = "auto",
  ...
)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{...}{Further arguments to be passed to \code{\link{BFA}()}.}
}
\description{
This function defines a BFA module for incorporation into a
psychTestR timeline.
Use this function if you want to include the Aesthetic Reponsiveness Assessment (AReA) in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the BFA,
consider using \code{\link{BFA_standalone}()}.
}
