% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{BFA_standalone}
\alias{BFA_standalone}
\title{BFA Standalone}
\usage{
BFA_standalone(languages = mpipoet::languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), and \code{"de"} (German).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{BFA_standalone}()}.}
}
\description{
This function launches a standalone testing session for the BFA questionnaire.
BFA stands for 'Big Five Aspects Scale'.
}
