% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{BFI_standalone}
\alias{BFI_standalone}
\title{BFI Standalone}
\usage{
BFI_standalone(languages = mpipoet::languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English),  \code{"de"} (informal German), and \code{"de_f"} (formal German)).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{BFI_standalone}()}.}
}
\description{
This function launches a standalone testing session for the ARA questionnaire.
BFI stands for 'Big Five Inventory'.
}
