% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIQ.R
\name{LIQ}
\alias{LIQ}
\title{LIQ}
\usage{
LIQ(
  label = "LIQ",
  with_intro = TRUE,
  with_finish = FALSE,
  dict = mpipoet::mpipoet_dict
)
}
\arguments{
\item{label}{(Character scalar) Label to give the LIQ results in the output file.}

\item{with_intro}{(Logical scalar) Flag whether to show intro page or not.}

\item{with_finish}{(Logical scalar) Flag whether to show final page or not.}

\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
This function defines a Literature Sophistication Index module for incorporation into a
psychTestR timeline.
Use this function if you want to include the LIQ in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the LIQ,
consider using \code{\link{LIQ_standalone}()}.
}
