% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIQ.R
\name{LIQ_standalone}
\alias{LIQ_standalone}
\title{Standalone LIQ}
\usage{
LIQ_standalone(
  title = NULL,
  with_id = FALSE,
  admin_password = "conifer",
  researcher_email = "klaus.frieler@ae.mpg.de",
  languages = c("en", "de"),
  dict = mpipoet::mpipoet_dict,
  validate_id = "auto",
  ...
)
}
\arguments{
\item{title}{(Scalar character) Title to display during testing.}

\item{with_id}{(Logical scalar) Whether to show a ID page.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}),
and German (\code{"DE"}).
The first language is selected by default}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{validate_id}{(Character scalar or closure) Function for validating IDs or string "auto" for default validation
which means ID should consist only of  alphanumeric characters.}

\item{...}{Further arguments to be passed to \code{\link{ART}()}.}
}
\description{
This function launches a standalone testing session for the LIQ (Literature Sophistication Index)
This can be used for data collection, either in the laboratory or online.
}
