% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIQ.R
\name{LSI}
\alias{LSI}
\title{LSI}
\usage{
LSI(label = "LSI", with_finish = FALSE, dict = mpipoet::mpipoet_dict, ...)
}
\arguments{
\item{label}{(Character scalar) Label to give the LSI results in the output file.}

\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
This function defines a Literature Sophistication Index module for incorporation into a
psychTestR timeline.
Use this function if you want to include the LSI in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the LSI,
consider using \code{\link{LSI_standalone}()}.
}
