% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLS.R
\name{SLS}
\alias{SLS}
\title{SLS}
\usage{
SLS(
  num_items = 36L,
  with_welcome = TRUE,
  with_training = FALSE,
  with_finish = TRUE,
  with_feedback = FALSE,
  label = "SLS",
  dict = mpipoet::mpipoet_dict,
  timeout = 10,
  ...
)
}
\arguments{
\item{num_items}{(Integer scalar) Number of items in the test.}

\item{with_welcome}{(Logical scalar) Whether to show a welcome page.}

\item{with_training}{(Logical scalar) Whether to include the training phase.}

\item{with_finish}{(Logical scalar) Whether to show a finished page.}

\item{with_feedback}{(Logical scalar) Whether to include feedback to the participants.}

\item{label}{(Character scalar) Label to give the SLS results in the output file.}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{timeout}{(Double scalar) The time to answer (in seconds)}

\item{...}{Further arguments to be passed to \code{SLS_main_test()}.}
}
\description{
This function defines a SLS  module for incorporation into a
psychTestR timeline.
Use this function if you want to include the MSM in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For demoing the MSM, consider using \code{\link{SLS_demo}()}.
For a standalone implementation of the SLS,
consider using \code{\link{SLS_standalone}()}.
}
