% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLS.R
\name{SLS_demo}
\alias{SLS_demo}
\title{Demo SLS}
\usage{
SLS_demo(
  num_items = 3L,
  title = "SLS Demo",
  dict = mpipoet::mpipoet_dict,
  admin_password = "demo",
  researcher_email = "klaus.frieler@ae.mpg.de",
  language = c("en", "de", "de_f")
)
}
\arguments{
\item{num_items}{(Integer scalar) Number of items in the test.}

\item{title}{(Character scalar) The title}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.
Defaults to \code{"demo"}.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.
Defaults to \email{longgold@gold.uc.ak},
the email address of this package's developer.}

\item{language}{The language you want to run your demo in.
Possible languages include English (\code{"en"}) and German (informal: \code{"DE"} and formal: \code{"DE_F"} )..
The first language is selected by default}

\item{...}{Further arguments to be passed to \code{\link{SLS}()}.}
}
\description{
This function launches a demo for the SLS
}
